Submit an expression to be evaluated to multiple jobs.

Submit an expression to be evaluated to multiple jobs.

Slurm_EvalQ(expr, njobs = 2L, job_name = opts_sluRm$get_job_name(),
  tmp_path = opts_sluRm$get_tmp_path(), plan = "collect",
  sbatch_opt = list(ntasks = 1L), rscript_opt = list(vanilla = TRUE),
  seeds = NULL, compress = TRUE, export = NULL,
  libPaths = .libPaths(), hooks = NULL)

Arguments

expr
An expression to be passed to Slurm.

njobs
Integer. Number of jobs to use in the job-array.

job_name
Character. Name of the job to be passed to Slurm.

tmp_path
Character. Path to the directory where all the data (including scripts) will be stored. Notice that this path must be accessible by all the nodes in the network (See opts_sluRm).

plan
A character scalar. (See the_plan).

sbatch_opt, rscript_opt
List. Options to be passed via flags to the bash file as #SBATCH and to Rscript respectively.

seeds
Integer vector of length njobs. Seeds to be passed to each job. When NULL (default), seeds will be picked automatically (see new_rscript).

compress
Logical scalar (default TRUE). Passed to saveRDS. Setting this value to FALSE can be useful when the user requires faster read/write of R objects on disk.

export
A named list with objects to be included in the Spawned sessions.

libPaths
A character vector. See .libPaths.

hooks
A list of functions (passed to new_slurm_job).